update types for portability
This is a long shot (pun intended)...

We want to be able to build and run on 32-bit systems. Having
`int64t`s creates a build & compatibility issue. However, `long
long` is compiler implementation specific, so although it does
guarantee that the type is "at least 64-bit wide", it seems that
it is not the correct way to go.

We must keep this in mind if something breaks down the road.

+++ examples/tf_inference.c
@@ -44,7 +44,7 @@ int main(int argc, char *argv[])
 		exit(1);
 	}
 
-	printf("Created new model %ld\n", vaccel_tf_saved_model_id(&model));
+	printf("Created new model %lld\n", vaccel_tf_saved_model_id(&model));
 	
 	ret = vaccel_sess_init(&vsess, 0);
 	if (ret) {
@@ -72,7 +72,7 @@ int main(int argc, char *argv[])
 	/* Input tensors & nodes */
 	struct vaccel_tf_node in_node = { "serving_default_input_1", 0 };
 	
-	int64_t dims[] = {1, 30};
+	long long int dims[] = {1, 30};
 	float data[30];
 	for (int i = 0; i < 30; ++i)
 		data[i] = 1.00;
@@ -102,7 +102,7 @@ int main(int argc, char *argv[])
 	printf("Success!\n");
 	printf("Output tensor => type:%u nr_dims:%u\n", out->data_type, out->nr_dims);
 	for (int i = 0; i < out->nr_dims; ++i)
-		printf("dim[%d]: %ld\n", i, out->dims[i]);
+		printf("dim[%d]: %lld\n", i, out->dims[i]);
 		
 	printf("Result Tensor :\n");
 	float *offsets = (float *)out->data;
